+++ src/sw/allotmentclub/browser/login.py
@@ -1,11 +1,10 @@
 # coding:utf-8
 from __future__ import unicode_literals
 from ..log import auth_log, log_with_user
-from ..user import User
+from .. import User, DashboardData
 from datetime import datetime
 from pyramid.security import NO_PERMISSION_REQUIRED
 from pyramid.view import view_config, forbidden_view_config
-import lnetatmo
 import hashlib
 import pyramid.httpexceptions
 import pyramid.security
@@ -71,43 +70,15 @@ class LoginView(sw.allotmentclub.browser.base.View):
                     gravatar=gravatar_url)
 
     def get_temp_data(self):
-        result = dict()
-        settings = pyramid.threadlocal.get_current_registry().settings
-        # 1 : Authenticate
-        if not settings.get('netatmo.client.id'):
+        data = (DashboardData.query()
+                .filter(DashboardData.rotersee_out_temp.isnot(None))
+                .order_by(DashboardData.date.desc()).first())
+        if data is None:
             return dict()
-        try:
-            authorization = lnetatmo.ClientAuth(
-                settings['netatmo.client.id'],
-                settings['netatmo.client.secret'],
-                settings['netatmo.user'],
-                settings['netatmo.pass'],
-                scope="read_station")
-        except Exception:
-            return dict()
-
-        # 2 : Get devices list
-        weatherData = lnetatmo.WeatherStationData(authorization)
-
-        # 3 : Access most fresh data directly
-        for module in weatherData.stationByName('Roter See')['modules']:
-            if module['type'] == 'NAModule1':
-                temp_trend = module['dashboard_data'].get(
-                    'temp_trend', 'stable')
-                temp_trend = ('right' if temp_trend == 'stable'
-                              else temp_trend)
-                result['temperature'] = (
-                    module['dashboard_data'].get('Temperature', '--'))
-                result['trend'] = temp_trend
-            elif module['type'] == 'NAModule3':
-                result['rain'] = module['dashboard_data'].get('Rain', '--')
-                result['sum_rain_1'] = (
-                    module['dashboard_data'].get('sum_rain_1', '--'))
-                result['sum_rain_24'] = (
-                    module['dashboard_data'].get('sum_rain_24', '--'))
-        result['date'] = datetime.fromtimestamp(
-            module['last_message']).strftime('%d.%m.%Y %H:%M Uhr')
-        return result
+        return dict(temperature=data.rotersee_out_temp,
+                    trend=data.rotersee_out_temp_trend,
+                    date=data.date.strftime('%d.%m.%Y %H:%M Uhr'),
+                    sum_rain_24=data.wachtelberg_rain_24)
 
     def __call__(self):
         result = self.login()